# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCUMENT RWANDA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/register/sendotp": {
            "post": {
                "tags": [
                    "Register"
                ],
                "description": "Renvoyer OTP",
                "operationId": "sendotp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "lg"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "fabio@yopmail.com"
                                    },
                                    "lg": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Renvoyer OTP."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/register/validotp": {
            "post": {
                "tags": [
                    "Register"
                ],
                "description": "Validation du Code OTP",
                "operationId": "validotp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "lg"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    },
                                    "lg": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation du Code OTP."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/register/forms": {
            "post": {
                "tags": [
                    "Register"
                ],
                "description": "Account creation",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "lg",
                                    "lastname",
                                    "firstname",
                                    "gender",
                                    "number",
                                    "email",
                                    "birthday",
                                    "birthplace",
                                    "profession",
                                    "village",
                                    "street_number",
                                    "house_number",
                                    "family_number",
                                    "fullname_person",
                                    "number_person",
                                    "residence_person",
                                    "cellule_id",
                                    "maritalstatus_id",
                                    "town_id",
                                    "fullname_father",
                                    "fullname_mother",
                                    "photo"
                                ],
                                "properties": {
                                    "lg": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "birthplace": {
                                        "type": "string"
                                    },
                                    "profession": {
                                        "type": "string"
                                    },
                                    "village": {
                                        "type": "string"
                                    },
                                    "street_number": {
                                        "type": "string"
                                    },
                                    "house_number": {
                                        "type": "string"
                                    },
                                    "family_number": {
                                        "type": "integer"
                                    },
                                    "fullname_person": {
                                        "type": "string"
                                    },
                                    "number_person": {
                                        "type": "string"
                                    },
                                    "residence_person": {
                                        "type": "string"
                                    },
                                    "cellule_id": {
                                        "type": "integer"
                                    },
                                    "town_id": {
                                        "type": "integer"
                                    },
                                    "bp": {
                                        "type": "string"
                                    },
                                    "diplome": {
                                        "type": "string"
                                    },
                                    "distinction": {
                                        "type": "string"
                                    },
                                    "maritalstatus_id": {
                                        "type": "integer"
                                    },
                                    "nationality_id": {
                                        "type": "integer"
                                    },
                                    "fullname_father": {
                                        "type": "string"
                                    },
                                    "fullname_mother": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Création de compte éffectuée avec succès."
                    },
                    "401": {
                        "description": "Echec de Création de compte."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
...
    },
...
}